refactor: move internal hooks to an internal folder
